Move json-to-outputs and generate-context to generate-outputs module, fix readme and use configuration checks in __main__.py
