+++ demisto_sdk/commands/generate_outputs/generate_context/tests/test_data/fake_examples.txt
